% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnergy.R
\name{getEnergy}
\alias{getEnergy}
\title{Get the Accumulated Cyclone Energy of a Given Storm}
\usage{
getEnergy(id)
}
\arguments{
\item{id}{storm id found in hurdat$id}
}
\value{
a numeric representing the ACE of a given storm
}
\description{
Given the storm id from the hurdat dataset, return the accumulated cyclone energy (ACE)
defined in https://en.wikipedia.org/wiki/Accumulated_cyclone_energy.
The formula is ACE = 1e-4 * sum(maximum velocities^2), where the
maximum velocities are estimated sustained wind speed in knots
measured at six-hour intervals
}
\examples{
getEnergy("AL101869")
}
