% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madeLandfall.R
\name{madeLandfall}
\alias{madeLandfall}
\title{Determine whether a given storm made landfall in the U.S.}
\usage{
madeLandfall(id)
}
\arguments{
\item{id}{storm id found in hurdat$id}
}
\value{
a boolean indicating whether the storm made landfall
}
\description{
Given the storm id from the hurdat dataset, return a boolean indicating whether or not
the storm made landfall, by checking whether the coordinates in the hurdat dataset
is in the U.S. boundaries as given by the maps package.
}
\examples{
#check if Hurricane Katrina (2005) made landfall
madeLandfall("AL122005")

}
