% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStorm.R
\name{plotStorm}
\alias{plotStorm}
\title{Plot a hurricane from the hurdat dataset}
\usage{
plotStorm(rowHurdat)
}
\arguments{
\item{row}{of hurdat database (indexed directly from hurdat)}
}
\value{
a map containing a graph of the hurricane from the row
}
\description{
Given a row from the hurdat dataset,
return a map of the United States which
contains county and state boundaries and the potential
hurricane view based off the maximum radii 
for each wind speed
}
\details{
The graph was calculated based off of 
an elliptical arc with the radii as reference
}
\examples{
#plot Hurricane Katrina around when it makes landfall
plotStorm(hurdat[c("46347"),])

}
