% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStormTracks.R
\name{plotStormTracks}
\alias{plotStormTracks}
\title{Plot a map of storm tracks for a selection of storms}
\usage{
plotStormTracks(storm_id)
}
\arguments{
\item{id}{storm id found in hurdat$id}
}
\value{
a map containing storm tracks
}
\description{
Given a vector of storm id from the hurdat dataset,
return a map of the United States which
contains county and state boundaries and the storm track of the storms
in the given vector.
}
\examples{
map <- plotStormTracks(c("AL051896", "AL011851"))
print(map)
}
