% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormTrack.R
\name{stormTrack}
\alias{stormTrack}
\title{Interpolate Actual Latitude and Longitude Data into 30-min Intervals}
\usage{
stormTrack(id)
}
\arguments{
\item{id}{storm id found in hurdat$id}
}
\value{
a matrix containing the following columns
\itemize{
  \item \code{@context} metadata
  \item \code{intervals} a vector of the time stamps in 30-min intervals in POSIXct
  \item \code{latitude} latitude in decimal degrees north (positive for northern hemisphere)
  \item \code{longitude} longitude in decimal degrees east (positive for eastern hemisphere)
}
}
\description{
Given the storm id from the hurdat dataset, return a matrix containing information
about the interpolated latitude and longitude values of that storm, where the
interpolation takes place in 30 minute intervals between the start and end time
of the storm.
}
\examples{
track <- stormTrack("AL041932")
print(track)
}
